refactor: split oversized modules under S104 and clear SonarCloud findings
