+++ src/aptl/backends/raes_participant_bindings.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from collections.abc import Callable, Mapping
 from dataclasses import dataclass
-from pathlib import Path
 from raes_contracts.planning import ProvisioningPlan
 
 from aptl.backends.raes_profiles import (
